refactor(ai-assistant): consolidate user-role-id resolution into one shared helper
The mcp-key, session-key, and chat routes each carried a verbatim copy of
getUserRoleIds, and the new mcp:ensure-api-key lib had become the fourth.
All four now use lib/user-role-ids.ts, which is also properly typed (drops
the `as any` casts the copies carried), so every surface that mints a
role-carrying API key inherits the same ACL from the same code path.

Closes the deferred finding from the pre-PR code review.
