+++ .ai/specs/2026-07-07-windows-one-command-agentic-dev-environment.md
@@ -131,3 +131,4 @@ The `ai-assistant` AGENTS.md "Ask First" gate on OpenCode Docker configuration /
 - **2026-07-07** — Initial implementation: containerized MCP service + full three-way wiring in both fullapp compose files, `mcp:ensure-api-key` CLI (+6 unit tests), OpenCode key-file delivery, `start-windows.bat`/`stop-windows.bat`/`scripts/windows/start-dev.ps1` one-command launcher, `.gitattributes` CRLF rules, `.env.example` + `docker/README.md` + installation docs updates.
 - **2026-07-07** — Rebased onto `feat/agent-orchestrator-mvp`: preserved that branch's OpenRouter provider support, file-defined agents/skills bind mounts on the `opencode` service, `OPENCODE_CONFIG_DIR` override, and the app-service `OPENCODE_URL` it already added in prod compose (deduplicated); the `.bat`/`.ps1` CRLF rules are now pattern-specific exceptions under the branch's global `* text=auto eol=lf` policy (#3470). Also fixed a leftover rebase conflict marker (`>>>>>>> 887ef17df`) that shipped inside `docker-compose.fullapp.dev.yml` on that branch and made the file unparseable for compose. The branch's `opencodeSessionId` API-key binding (cross-user session-continuation fix) is orthogonal to the MCP server key and required no changes.
 - **2026-07-07** — Post-review hardening (10-angle code review): owner lookup made genuinely encrypted-email-aware (`$or` emailHash pattern from `AuthService.findUserByEmail` — a plaintext `{ email }` filter finds nothing under `TENANT_DATA_ENCRYPTION=true`); key file written 0644 (root-owned 0600 was unreadable by the non-root OpenCode container); stale-key cleanup tenant-scoped; bare `--file` flag no longer provisions into a file named "true"; canonical env var unified to `MCP_SERVER_API_KEY_FILE`; OpenCode entrypoint degrades to headerless start (with loud warnings) instead of crash-looping on timeout/read failure, bounds each probe with `--max-time`, and tolerates trailing slashes in `OPENCODE_MCP_URL`; `mcpKeyConfigured` validates the file secret against the DB; app services gained `restart: unless-stopped` (reboot-survival parity with mcp/opencode); dev `mcp` service reuses the app-built image without a duplicate build block; healthchecks switched to busybox `wget`/bounded `curl` at 30s and documented as visibility-only. Launcher: compose output now streams to the console (`-Status`/`-Logs`/failure diagnostics were swallowed), `-DryRun` is forwarded on standalone re-entry and exits cleanly pre-clone, `.env` writes are BOM-free with LF endings + trailing-newline guard, empty `KEY=` placeholders are fillable, host ports resolve from `.env` instead of hardcoded literals, `wsl --status` replaces the stub-fooled existence probe, `compose ps --format json` parsing tolerates array and NDJSON output, Defender/docker-users probe failures no longer re-trigger UAC or reboots every run, Git's PATH is refreshed after elevated install (same-run clone), the transcript stops before the summary so the superadmin password never lands in the %TEMP% log, OM_AI_MODEL is no longer pinned into `.env`, and the write-only state file + `-Resumed` plumbing were dropped. `.bat` files are committed with literal CRLF bytes and marked `-text` (raw GitHub downloads and ZIPs now get CRLF too). `packages/ai-assistant/AGENTS.md` two-tier-auth guidance updated for file-based key delivery.
+- **2026-07-08** — Closed the review's deferred finding: the four verbatim copies of the user-role-id resolution (mcp-key, session-key, and chat routes plus the CLI provisioning lib) now share one typed helper, `lib/user-role-ids.ts` (`getUserRoleIds`), which also drops the `as any` casts the copies carried — every role-carrying API key surface now inherits the same ACL from the same code path.
